% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepspecify-package.r
\docType{package}
\name{sweepspecify}
\alias{sweepspecify}
\alias{sweepspecify-package}
\title{Sweep & Specify Operations for the OUHSC CDW}
\description{
Thanks to Funders, ...
}
\details{
sweepspecify
}
\note{
The most recent development version is available through \href{https://github.com/OuhscBbmc/sweepspecify}{GitHub} by
running
\code{devtools::install_github} \code{(repo = 'OuhscBbmc/sweepspecify')}
(make sure \href{https://cran.r-project.org/package=devtools}{devtools} is already installed).
If you're having trouble with the package, please install the development version.  If this doesn't solve
your problem, please create a \href{https://github.com/OuhscBbmc/sweepspecify/issues}{new issue}, or email Will.
}
\examples{
\dontrun{
# Install/update sweepspecify with the release version from CRAN.
install.packages('sweepspecify') #But it's not on CRAN yet.

# Install/update sweepspecify with the development version from GitHub
#install.packages('devtools') #Uncomment if `devtools` isn't installed already.
devtools::install_github('OuhscBbmc/sweepspecify')
}
}
\author{
\href{http://scholar.google.com/citations?user=ffsJTC0AAAAJ}{William Howard Beasley},
University of Oklahoma Health Sciences Center, College of Medicine,
Dept of Pediatrics, \href{http://www.ouhsc.edu/bbmc/#team}{BBMC}.

Maintainer: Will Beasley \href{mailto:wibeasley@hotmail.com}{wibeasley@hotmail.com}
}
